[droid-config-sailfish] Adjust sailfish-upgrade-ui's scaling factor settings
Both UI resources (z1.5 and z1.75) are probably going to be installed, and the
latter, being later in the order, will override the correct settings for the
X Compact's screen.

This commit introduces a symlink to the correct configuration file. This link
would be picked up last by systemd so that eventually the correct settings
are always picked up.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>