[hybris-libsensorfw-qt5] Added a compat package that disables some secondary sensors
Often the sensorfw daemon seems to enter an unrecovable state where it begins to
eat all the available CPU until the process is stopped.

Forcibly disabling the secondary sensors has helped. Further testing is needed
to isolate the actual sensor that is misbehaving.

Note that the sensors are disabled via a diversion. The sensorfw daemon in future
Sailfish OS releases will probably allow to disable the sensor from a configuration
file.

The disabled sensors are:
  * gyroscope
  * magnetometer
  * pressure
  * step counter

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>